test: Slider 스토리 추가
